\name{About this package}
\alias{About this package}
\title{kmetricsパッケージについて}
\usage{
}
\description{
RStudio v1.3 のチュートリアル機能（in-IDE Tutorial）にチュートリアルファイルを追加するパッケージです。
}
\examples{}
\details{
本パッケージにはチュートリアルファイルとヘルプ以外は含まれていません。

各チュートリアルファイルで必要となるパッケージがあります。チュートリアルが実行できない場合はヘルプを参照して必要なパッケージをインストールしてください。
}
