\name{Time Series}
\alias{Time Series}
\title{時系列データ分析入門のために}
\usage{
click "Start Tutorial" in "Tutorial" pane.
}
\description{
データ分析勉強会では取り上げていない時系列型（`ts`クラス）の入門用チュートリアルです。
}
\examples{}
\details{
実行には `forecast` パッケージが必要です。

インストールしていない場合はコンソールから以下のコードを実行し `forecast` パッケージをインストールしてからチュートリアルを実行してください。

\code{install.packages("forecast")}
}
