\name{Time Series}
\alias{Time Series}
\title{時系列データ分析入門のためのチュートリアル}
\usage{
Click "Start Tutorial" in "Tutorial" pane.
}
\description{
データ分析勉強会では取り上げていない時系列型（`ts`クラス）の入門用チュートリアルです。
}
\examples{}
\details{
本チュートリアルの実行には`forecast`パッケージが必要です。

`forecast`パッケージがインストールされていない場合にはコンソールから以下のコードを実行しパッケージをインストールしてください。

\code{install.packages("forecast")}
}
